fix(auth): rename proxy.ts to middleware.ts with correct export name
Next.js only picks up middleware from a file named `middleware.ts`.
The previous file was named `proxy.ts` and exported a function named
`proxy` instead of `middleware`, so the session refresh never ran.

Agent-Logs-Url: https://github.com/alexisbohns/pbbls/sessions/65aa67ff-b7da-4dc2-b1f5-9c1d6d91ba12
