+++ apps/web/middleware.ts
@@ -1,7 +1,7 @@
 import { type NextRequest } from "next/server"
-import { updateSession } from "@/lib/supabase/proxy"
+import { updateSession } from "@/lib/supabase/middleware"
 
-export async function proxy(request: NextRequest) {
+export async function middleware(request: NextRequest) {
   return await updateSession(request)
 }
